split up 'processing.py' into subscripts
